LPS-27364
