+++ portal-web/test/com/liferay/portalweb/portlet/documentsandmediadisplay/dmdocument/deletedmdocumentsdmd/DeleteDMDocumentsDMDTest.html
@@ -21,6 +21,11 @@
 	<td>Documents and Media Test Page</td>
 </tr>
 <tr>
+	<td>pause</td>
+	<td>5000</td>
+	<td></td>
+</tr>
+<tr>
 	<td>storeElementPresent</td>
 	<td>//span[@class='entry-title']</td>
 	<td>dmDocument1Present</td>
@@ -69,6 +74,16 @@
 	<td>assertText</td>
 	<td>//div[@class='portlet-msg-success']</td>
 	<td>Your request completed successfully.</td>
+</tr>
+<tr>
+	<td>label</td>
+	<td>DMDocument1NotPresent</td>
+	<td></td>
+</tr>
+<tr>
+	<td>pause</td>
+	<td>5000</td>
+	<td></td>
 </tr>
 <tr>
 	<td>storeElementPresent</td>
@@ -121,6 +136,16 @@
 	<td>Your request completed successfully.</td>
 </tr>
 <tr>
+	<td>label</td>
+	<td>DMDocument2NotPresent</td>
+	<td></td>
+</tr>
+<tr>
+	<td>pause</td>
+	<td>5000</td>
+	<td></td>
+</tr>
+<tr>
 	<td>storeElementPresent</td>
 	<td>//span[@class='entry-title']</td>
 	<td>dmDocument3Present</td>
@@ -129,11 +154,6 @@
 	<td>gotoIf</td>
 	<td>!${dmDocument3Present}</td>
 	<td>DMDocument3NotPresent</td>
-</tr>
-<tr>
-	<td>pause</td>
-	<td>5000</td>
-	<td></td>
 </tr>
 <tr>
 	<td>assertText</td>
@@ -174,16 +194,6 @@
 	<td>assertText</td>
 	<td>//div[@class='portlet-msg-success']</td>
 	<td>Your request completed successfully.</td>
-</tr>
-<tr>
-	<td>label</td>
-	<td>DMDocument1NotPresent</td>
-	<td></td>
-</tr>
-<tr>
-	<td>label</td>
-	<td>DMDocument2NotPresent</td>
-	<td></td>
 </tr>
 <tr>
 	<td>label</td>
