+++ portal-web/test/com/liferay/portalweb/portlet/documentsandmediadisplay/dmdocument/deletedmdocumentsdmd/DeleteDMDocumentsDMDTests.html
@@ -10,14 +10,14 @@
 <tr><td><b>DeleteDMDocumentsDMDTests</b></td></tr>
 <tr><td><a href="../../../../../portalweb/portlet/documentsandmediadisplay/portlet/addportletdmd/AddPageDMDTest.html">AddPageDMDTest</a></td></tr>
 <tr><td><a href="../../../../../portalweb/portlet/documentsandmediadisplay/portlet/addportletdmd/AddPortletDMDTest.html">AddPortletDMDTest</a></td></tr>
-<tr><td><a href="../../../../../portalweb/portlet/documentsandmediadisplay/portlet/configureportletdmdshowfoldermenu/ConfigurePortletShowFolderMenuTest.html">ConfigurePortletShowFolderMenuTest</a></td></tr>
-<tr><td><a href="../../../../../portalweb/portlet/documentsandmediadisplay/portlet/configureportletdmdshowactions/ConfigurePortletShowActionsTest.html">ConfigurePortletShowActionsTest</a></td></tr>
-<tr><td><a href="../../../../../portalweb/portlet/documentsandmediadisplay/portlet/configureportletdmdshownavigationlinks/ConfigurePortletShowNavigationLinksTest.html">ConfigurePortletShowNavigationLinksTest</a></td></tr>
-<tr><td><a href="../../../../../portalweb/portlet/documentsandmediadisplay/portlet/configureportletdmdshowsearch/ConfigurePortletShowSearchTest.html">ConfigurePortletShowSearchTest</a></td></tr>
-<tr><td><a href="../../../../../portalweb/portlet/documentsandmediadisplay/portlet/configureportletdmdshowfoldermenu/ViewPortletShowFolderMenuTest.html">ViewPortletShowFolderMenuTest</a></td></tr>
-<tr><td><a href="../../../../../portalweb/portlet/documentsandmediadisplay/portlet/configureportletdmdshowactions/ViewPortletShowActionsTest.html">ViewPortletShowActionsTest</a></td></tr>
-<tr><td><a href="../../../../../portalweb/portlet/documentsandmediadisplay/portlet/configureportletdmdshownavigationlinks/ViewPortletShowNavigationLinksTest.html">ViewPortletShowNavigationLinksTest</a></td></tr>
-<tr><td><a href="../../../../../portalweb/portlet/documentsandmediadisplay/portlet/configureportletdmdshowsearch/ViewPortletShowSearchTest.html">ViewPortletShowSearchTest</a></td></tr>
+<tr><td><a href="../../../../../portalweb/portlet/documentsandmediadisplay/portlet/configureportletdmdshowfoldermenu/ConfigurePortletDMDShowFolderMenuTest.html">ConfigurePortletDMDShowFolderMenuTest</a></td></tr>
+<tr><td><a href="../../../../../portalweb/portlet/documentsandmediadisplay/portlet/configureportletdmdshowactions/ConfigurePortletDMDShowActionsTest.html">ConfigurePortletDMDShowActionsTest</a></td></tr>
+<tr><td><a href="../../../../../portalweb/portlet/documentsandmediadisplay/portlet/configureportletdmdshownavigationlinks/ConfigurePortletDMDShowNavigationLinksTest.html">ConfigurePortletDMDShowNavigationLinksTest</a></td></tr>
+<tr><td><a href="../../../../../portalweb/portlet/documentsandmediadisplay/portlet/configureportletdmdshowsearch/ConfigurePortletDMDShowSearchTest.html">ConfigurePortletDMDShowSearchTest</a></td></tr>
+<tr><td><a href="../../../../../portalweb/portlet/documentsandmediadisplay/portlet/configureportletdmdshowfoldermenu/ViewPortletDMDShowFolderMenuTest.html">ViewPortletDMDShowFolderMenuTest</a></td></tr>
+<tr><td><a href="../../../../../portalweb/portlet/documentsandmediadisplay/portlet/configureportletdmdshowactions/ViewPortletDMDShowActionsTest.html">ViewPortletDMDShowActionsTest</a></td></tr>
+<tr><td><a href="../../../../../portalweb/portlet/documentsandmediadisplay/portlet/configureportletdmdshownavigationlinks/ViewPortletDMDShowNavigationLinksTest.html">ViewPortletDMDShowNavigationLinksTest</a></td></tr>
+<tr><td><a href="../../../../../portalweb/portlet/documentsandmediadisplay/portlet/configureportletdmdshowsearch/ViewPortletDMDShowSearchTest.html">ViewPortletDMDShowSearchTest</a></td></tr>
 <tr><td><a href="../../../../../portalweb/portlet/documentsandmediadisplay/dmdocument/adddmdocumentdmd/AddDMDocument1DMDTest.html">AddDMDocument1DMDTest</a></td></tr>
 <tr><td><a href="../../../../../portalweb/portlet/documentsandmediadisplay/dmdocument/adddmdocumentdmd/AddDMDocument2DMDTest.html">AddDMDocument2DMDTest</a></td></tr>
 <tr><td><a href="../../../../../portalweb/portlet/documentsandmediadisplay/dmdocument/adddmdocumentdmd/AddDMDocument3DMDTest.html">AddDMDocument3DMDTest</a></td></tr>
