+++ portal-web/test/com/liferay/portalweb/portlet/documentsandmediadisplay/dmfolder/adddmfolderdmd/AddDMFolderDMDTests.java
@@ -18,14 +18,14 @@ import com.liferay.portalweb.portal.BaseTestSuite;
 import com.liferay.portalweb.portal.util.TearDownPageTest;
 import com.liferay.portalweb.portlet.documentsandmediadisplay.portlet.addportletdmd.AddPageDMDTest;
 import com.liferay.portalweb.portlet.documentsandmediadisplay.portlet.addportletdmd.AddPortletDMDTest;
-import com.liferay.portalweb.portlet.documentsandmediadisplay.portlet.configureportletdmdshowactions.ConfigurePortletShowActionsTest;
-import com.liferay.portalweb.portlet.documentsandmediadisplay.portlet.configureportletdmdshowactions.ViewPortletShowActionsTest;
-import com.liferay.portalweb.portlet.documentsandmediadisplay.portlet.configureportletdmdshowfoldermenu.ConfigurePortletShowFolderMenuTest;
-import com.liferay.portalweb.portlet.documentsandmediadisplay.portlet.configureportletdmdshowfoldermenu.ViewPortletShowFolderMenuTest;
-import com.liferay.portalweb.portlet.documentsandmediadisplay.portlet.configureportletdmdshownavigationlinks.ConfigurePortletShowNavigationLinksTest;
-import com.liferay.portalweb.portlet.documentsandmediadisplay.portlet.configureportletdmdshownavigationlinks.ViewPortletShowNavigationLinksTest;
-import com.liferay.portalweb.portlet.documentsandmediadisplay.portlet.configureportletdmdshowsearch.ConfigurePortletShowSearchTest;
-import com.liferay.portalweb.portlet.documentsandmediadisplay.portlet.configureportletdmdshowsearch.ViewPortletShowSearchTest;
+import com.liferay.portalweb.portlet.documentsandmediadisplay.portlet.configureportletdmdshowactions.ConfigurePortletDMDShowActionsTest;
+import com.liferay.portalweb.portlet.documentsandmediadisplay.portlet.configureportletdmdshowactions.ViewPortletDMDShowActionsTest;
+import com.liferay.portalweb.portlet.documentsandmediadisplay.portlet.configureportletdmdshowfoldermenu.ConfigurePortletDMDShowFolderMenuTest;
+import com.liferay.portalweb.portlet.documentsandmediadisplay.portlet.configureportletdmdshowfoldermenu.ViewPortletDMDShowFolderMenuTest;
+import com.liferay.portalweb.portlet.documentsandmediadisplay.portlet.configureportletdmdshownavigationlinks.ConfigurePortletDMDShowNavigationLinksTest;
+import com.liferay.portalweb.portlet.documentsandmediadisplay.portlet.configureportletdmdshownavigationlinks.ViewPortletDMDShowNavigationLinksTest;
+import com.liferay.portalweb.portlet.documentsandmediadisplay.portlet.configureportletdmdshowsearch.ConfigurePortletDMDShowSearchTest;
+import com.liferay.portalweb.portlet.documentsandmediadisplay.portlet.configureportletdmdshowsearch.ViewPortletDMDShowSearchTest;
 
 import junit.framework.Test;
 import junit.framework.TestSuite;
@@ -38,14 +38,14 @@ public class AddDMFolderDMDTests extends BaseTestSuite {
 		TestSuite testSuite = new TestSuite();
 		testSuite.addTestSuite(AddPageDMDTest.class);
 		testSuite.addTestSuite(AddPortletDMDTest.class);
-		testSuite.addTestSuite(ConfigurePortletShowFolderMenuTest.class);
-		testSuite.addTestSuite(ConfigurePortletShowActionsTest.class);
-		testSuite.addTestSuite(ConfigurePortletShowNavigationLinksTest.class);
-		testSuite.addTestSuite(ConfigurePortletShowSearchTest.class);
-		testSuite.addTestSuite(ViewPortletShowFolderMenuTest.class);
-		testSuite.addTestSuite(ViewPortletShowActionsTest.class);
-		testSuite.addTestSuite(ViewPortletShowNavigationLinksTest.class);
-		testSuite.addTestSuite(ViewPortletShowSearchTest.class);
+		testSuite.addTestSuite(ConfigurePortletDMDShowFolderMenuTest.class);
+		testSuite.addTestSuite(ConfigurePortletDMDShowActionsTest.class);
+		testSuite.addTestSuite(ConfigurePortletDMDShowNavigationLinksTest.class);
+		testSuite.addTestSuite(ConfigurePortletDMDShowSearchTest.class);
+		testSuite.addTestSuite(ViewPortletDMDShowFolderMenuTest.class);
+		testSuite.addTestSuite(ViewPortletDMDShowActionsTest.class);
+		testSuite.addTestSuite(ViewPortletDMDShowNavigationLinksTest.class);
+		testSuite.addTestSuite(ViewPortletDMDShowSearchTest.class);
 		testSuite.addTestSuite(AddDMFolderDMDTest.class);
 		testSuite.addTestSuite(ViewDMFolderDMDTest.class);
 		testSuite.addTestSuite(TearDownDMDFolderTest.class);
