+++ portal-web/test/com/liferay/portalweb/portlet/documentsandmediadisplay/portlet/addportletsdmd/AddPortletsDMDTests.html
@@ -10,8 +10,6 @@
 <tr><td><b>AddPortletsDMDTests</b></td></tr>
 <tr><td><a href="../../../../../portalweb/portlet/documentsandmediadisplay/portlet/addportletdmd/AddPageDMDTest.html">AddPageDMDTest</a></td></tr>
 <tr><td><a href="../../../../../portalweb/portlet/documentsandmediadisplay/portlet/addportletdmd/AddPortletDMD1Test.html">AddPortletDMD1Test</a></td></tr>
-
-
 <tr><td><a href="../../../../../portalweb/portlet/documentsandmediadisplay/portlet/addportletdmd/AddPortletDMD2Test.html">AddPortletDMD2Test</a></td></tr>
 <tr><td><a href="../../../../../portalweb/portlet/documentsandmediadisplay/portlet/addportletdmd/AddPortletDMD3Test.html">AddPortletDMD3Test</a></td></tr>
 <tr><td><a href="ViewPortletsDMDTest.html">ViewPortletsDMDTest</a></td></tr>
