+++ portal-web/test/com/liferay/portalweb/portlet/documentsandmediadisplay/portlet/configureportletdmdshowactions/ConfigurePortletDMDShowActionsTests.html
@@ -3,16 +3,19 @@
 <html xmlns="http://www.w3.org/1999/xhtml" xml:lang="en" lang="en">
 <head>
   <meta content="text/html; charset=UTF-8" http-equiv="content-type" />
-  <title>ConfigurePortletShowFolderMenuTests</title>
+  <title>ConfigurePortletDMDShowActionsTests</title>
 </head>
 <body>
 <table id="suiteTable" cellpadding="1" cellspacing="1" border="1" class="selenium"><tbody>
-<tr><td><b>ConfigurePortletShowFolderMenuTests</b></td></tr>
+<tr><td><b>ConfigurePortletDMDShowActionsTests</b></td></tr>
 <tr><td><a href="../../../../../portalweb/portlet/documentsandmediadisplay/portlet/addportletdmd/AddPageDMDTest.html">AddPageDMDTest</a></td></tr>
 <tr><td><a href="../../../../../portalweb/portlet/documentsandmediadisplay/portlet/addportletdmd/AddPortletDMDTest.html">AddPortletDMDTest</a></td></tr>
-<tr><td><a href="ConfigurePortletShowFolderMenuTest.html">ConfigurePortletShowFolderMenuTest</a></td></tr>
-<tr><td><a href="ViewPortletShowFolderMenuTest.html">ViewPortletShowFolderMenuTest</a></td></tr>
+<tr><td><a href="../../../../../portalweb/portlet/documentsandmediadisplay/portlet/configureportletdmdshowfoldermenu/ConfigurePortletDMDShowFolderMenuTest.html">ConfigurePortletDMDShowFolderMenuTest</a></td></tr>
+<tr><td><a href="ConfigurePortletDMDShowActionsTest.html">ConfigurePortletDMDShowActionsTest</a></td></tr>
+<tr><td><a href="ViewPortletDMDShowActionsTest.html">ViewPortletDMDShowActionsTest</a></td></tr>
 <tr><td><a href="../../../../../portalweb/portal/util/TearDownPageTest.html">TearDownPageTest</a></td></tr>
 </tbody></table>
 </body>
 </html>
+
+
