+++ portal-web/test/com/liferay/portalweb/portlet/documentsandmediadisplay/portlet/configureportletdmdshowactions/ViewPortletDMDShowActionsTest.html
@@ -1,12 +1,12 @@
 <html>
 <head>
 <meta http-equiv="Content-Type" content="text/html; charset=UTF-8">
-<title>ViewPortletShowActionsTest</title>
+<title>ViewPortletDMDShowActionsTest</title>
 </head>
 <body>
 <table cellpadding="1" cellspacing="1" border="1">
 <thead>
-<tr><td rowspan="1" colspan="3">ViewPortletShowActionsTest</td></tr>
+<tr><td rowspan="1" colspan="3">ViewPortletDMDShowActionsTest</td></tr>
 </thead><tbody>
 <tr>
 	<td>open</td>
