+++ portal-web/test/com/liferay/portalweb/portlet/documentsandmediadisplay/portlet/configureportletdmdshowfoldermenu/ConfigurePortletDMDShowFolderMenuTest.html
@@ -1,12 +1,12 @@
 <html>
 <head>
 <meta http-equiv="Content-Type" content="text/html; charset=UTF-8">
-<title>ConfigurePortletShowNavigationLinksTest</title>
+<title>ConfigurePortletDMDShowFolderMenuTest</title>
 </head>
 <body>
 <table cellpadding="1" cellspacing="1" border="1">
 <thead>
-<tr><td rowspan="1" colspan="3">ConfigurePortletShowNavigationLinksTest</td></tr>
+<tr><td rowspan="1" colspan="3">ConfigurePortletDMDShowFolderMenuTest</td></tr>
 </thead><tbody>
 <tr>
 	<td>open</td>
@@ -60,23 +60,23 @@
 </tr>
 <tr>
 	<td>waitForVisible</td>
-	<td>//input[@id='_86_showTabsCheckbox']</td>
-	<td>Show Navigation Links</td>
+	<td>//input[@id='_86_showFolderMenuCheckbox']</td>
+	<td>Show Folder Menu</td>
 </tr>
 <tr>
 	<td>assertNotChecked</td>
-	<td>//input[@id='_86_showTabsCheckbox']</td>
-	<td>Show Navigation Links</td>
+	<td>//input[@id='_86_showFolderMenuCheckbox']</td>
+	<td>Show Folder Menu</td>
 </tr>
 <tr>
 	<td>clickAt</td>
-	<td>//input[@id='_86_showTabsCheckbox']</td>
-	<td>Show Navigation Links</td>
+	<td>//input[@id='_86_showFolderMenuCheckbox']</td>
+	<td>Show Folder Menu</td>
 </tr>
 <tr>
 	<td>assertChecked</td>
-	<td>//input[@id='_86_showTabsCheckbox']</td>
-	<td>Show Navigation Links</td>
+	<td>//input[@id='_86_showFolderMenuCheckbox']</td>
+	<td>Show Folder Menu</td>
 </tr>
 <tr>
 	<td>clickAtAndWait</td>
