+++ portal-web/test/com/liferay/portalweb/portlet/documentsandmediadisplay/portlet/configureportletdmdshowfoldermenu/ConfigurePortletDMDShowFolderMenuTests.html
@@ -3,15 +3,15 @@
 <html xmlns="http://www.w3.org/1999/xhtml" xml:lang="en" lang="en">
 <head>
   <meta content="text/html; charset=UTF-8" http-equiv="content-type" />
-  <title>ConfigurePortletShowNavigationLinksTests</title>
+  <title>ConfigurePortletDMDShowFolderMenuTests</title>
 </head>
 <body>
 <table id="suiteTable" cellpadding="1" cellspacing="1" border="1" class="selenium"><tbody>
-<tr><td><b>ConfigurePortletShowNavigationLinksTests</b></td></tr>
+<tr><td><b>ConfigurePortletDMDShowFolderMenuTests</b></td></tr>
 <tr><td><a href="../../../../../portalweb/portlet/documentsandmediadisplay/portlet/addportletdmd/AddPageDMDTest.html">AddPageDMDTest</a></td></tr>
 <tr><td><a href="../../../../../portalweb/portlet/documentsandmediadisplay/portlet/addportletdmd/AddPortletDMDTest.html">AddPortletDMDTest</a></td></tr>
-<tr><td><a href="ConfigurePortletShowNavigationLinksTest.html">ConfigurePortletShowNavigationLinksTest</a></td></tr>
-<tr><td><a href="ViewPortletShowNavigationLinksTest.html">ViewPortletShowNavigationLinksTest</a></td></tr>
+<tr><td><a href="ConfigurePortletDMDShowFolderMenuTest.html">ConfigurePortletDMDShowFolderMenuTest</a></td></tr>
+<tr><td><a href="ViewPortletDMDShowFolderMenuTest.html">ViewPortletDMDShowFolderMenuTest</a></td></tr>
 <tr><td><a href="../../../../../portalweb/portal/util/TearDownPageTest.html">TearDownPageTest</a></td></tr>
 </tbody></table>
 </body>
