+++ portal-web/test/com/liferay/portalweb/portlet/documentsandmediadisplay/portlet/configureportletdmdshowfoldermenu/ViewPortletDMDShowFolderMenuTest.html
@@ -1,12 +1,12 @@
 <html>
 <head>
 <meta http-equiv="Content-Type" content="text/html; charset=UTF-8">
-<title>ViewPortletShowNavigationLinksTest</title>
+<title>ViewPortletDMDShowFolderMenuTest</title>
 </head>
 <body>
 <table cellpadding="1" cellspacing="1" border="1">
 <thead>
-<tr><td rowspan="1" colspan="3">ViewPortletShowNavigationLinksTest</td></tr>
+<tr><td rowspan="1" colspan="3">ViewPortletDMDShowFolderMenuTest</td></tr>
 </thead><tbody>
 <tr>
 	<td>open</td>
@@ -25,8 +25,8 @@
 </tr>
 <tr>
 	<td>assertVisible</td>
-	<td>//ul[@class='top-links-navigation']</td>
-	<td>Navigation Links</td>
+	<td>//div[@class='lfr-asset-summary']</td>
+	<td>Folder Menu</td>
 </tr>
 </tbody></table>
 </body>
