+++ portal-web/test/com/liferay/portalweb/portlet/documentsandmediadisplay/portlet/configureportletdmdshownavigationlinks/ConfigurePortletDMDShowNavigationLinksTest.html
@@ -1,12 +1,12 @@
 <html>
 <head>
 <meta http-equiv="Content-Type" content="text/html; charset=UTF-8">
-<title>ConfigurePortletShowSearchTest</title>
+<title>ConfigurePortletDMDShowNavigationLinksTest</title>
 </head>
 <body>
 <table cellpadding="1" cellspacing="1" border="1">
 <thead>
-<tr><td rowspan="1" colspan="3">ConfigurePortletShowSearchTest</td></tr>
+<tr><td rowspan="1" colspan="3">ConfigurePortletDMDShowNavigationLinksTest</td></tr>
 </thead><tbody>
 <tr>
 	<td>open</td>
@@ -49,7 +49,7 @@
 	<td>Configuration</td>
 </tr>
 <tr>
-	<td>click</td>
+	<td>clickAt</td>
 	<td>//div[@class='lfr-component lfr-menu-list']/ul/li[2]/a</td>
 	<td>Configuration</td>
 </tr>
@@ -60,23 +60,23 @@
 </tr>
 <tr>
 	<td>waitForVisible</td>
-	<td>//input[@id='_86_showFoldersSearchCheckbox']</td>
-	<td>Show Search</td>
+	<td>//input[@id='_86_showTabsCheckbox']</td>
+	<td>Show Navigation Links</td>
 </tr>
 <tr>
 	<td>assertNotChecked</td>
-	<td>//input[@id='_86_showFoldersSearchCheckbox']</td>
-	<td>Show Search</td>
+	<td>//input[@id='_86_showTabsCheckbox']</td>
+	<td>Show Navigation Links</td>
 </tr>
 <tr>
 	<td>clickAt</td>
-	<td>//input[@id='_86_showFoldersSearchCheckbox']</td>
-	<td>Show Search</td>
+	<td>//input[@id='_86_showTabsCheckbox']</td>
+	<td>Show Navigation Links</td>
 </tr>
 <tr>
 	<td>assertChecked</td>
-	<td>//input[@id='_86_showFoldersSearchCheckbox']</td>
-	<td>Show Search</td>
+	<td>//input[@id='_86_showTabsCheckbox']</td>
+	<td>Show Navigation Links</td>
 </tr>
 <tr>
 	<td>clickAtAndWait</td>
