+++ portal-web/test/com/liferay/portalweb/portlet/documentsandmediadisplay/portlet/configureportletdmdshownavigationlinks/ConfigurePortletDMDShowNavigationLinksTests.html
@@ -3,15 +3,15 @@
 <html xmlns="http://www.w3.org/1999/xhtml" xml:lang="en" lang="en">
 <head>
   <meta content="text/html; charset=UTF-8" http-equiv="content-type" />
-  <title>ConfigurePortletShowSearchTests</title>
+  <title>ConfigurePortletDMDShowNavigationLinksTests</title>
 </head>
 <body>
 <table id="suiteTable" cellpadding="1" cellspacing="1" border="1" class="selenium"><tbody>
-<tr><td><b>ConfigurePortletShowSearchTests</b></td></tr>
+<tr><td><b>ConfigurePortletDMDShowNavigationLinksTests</b></td></tr>
 <tr><td><a href="../../../../../portalweb/portlet/documentsandmediadisplay/portlet/addportletdmd/AddPageDMDTest.html">AddPageDMDTest</a></td></tr>
 <tr><td><a href="../../../../../portalweb/portlet/documentsandmediadisplay/portlet/addportletdmd/AddPortletDMDTest.html">AddPortletDMDTest</a></td></tr>
-<tr><td><a href="ConfigurePortletShowSearchTest.html">ConfigurePortletShowSearchTest</a></td></tr>
-<tr><td><a href="ViewPortletShowSearchTest.html">ViewPortletShowSearchTest</a></td></tr>
+<tr><td><a href="ConfigurePortletDMDShowNavigationLinksTest.html">ConfigurePortletDMDShowNavigationLinksTest</a></td></tr>
+<tr><td><a href="ViewPortletDMDShowNavigationLinksTest.html">ViewPortletDMDShowNavigationLinksTest</a></td></tr>
 <tr><td><a href="../../../../../portalweb/portal/util/TearDownPageTest.html">TearDownPageTest</a></td></tr>
 </tbody></table>
 </body>
