+++ portal-web/test/com/liferay/portalweb/portlet/documentsandmediadisplay/portlet/configureportletdmdshownavigationlinks/ViewPortletDMDShowNavigationLinksTest.html
@@ -1,12 +1,12 @@
 <html>
 <head>
 <meta http-equiv="Content-Type" content="text/html; charset=UTF-8">
-<title>ViewPortletShowFolderMenuTest</title>
+<title>ViewPortletDMDShowNavigationLinksTest</title>
 </head>
 <body>
 <table cellpadding="1" cellspacing="1" border="1">
 <thead>
-<tr><td rowspan="1" colspan="3">ViewPortletShowFolderMenuTest</td></tr>
+<tr><td rowspan="1" colspan="3">ViewPortletDMDShowNavigationLinksTest</td></tr>
 </thead><tbody>
 <tr>
 	<td>open</td>
@@ -25,8 +25,8 @@
 </tr>
 <tr>
 	<td>assertVisible</td>
-	<td>//div[@class='lfr-asset-summary']</td>
-	<td>Folder Menu</td>
+	<td>//ul[@class='top-links-navigation']</td>
+	<td>Navigation Links</td>
 </tr>
 </tbody></table>
 </body>
