+++ portal-web/test/com/liferay/portalweb/portlet/documentsandmediadisplay/portlet/configureportletdmdshowsearch/ViewPortletDMDShowSearchTest.html
@@ -1,12 +1,12 @@
 <html>
 <head>
 <meta http-equiv="Content-Type" content="text/html; charset=UTF-8">
-<title>ViewPortletShowSearchTest</title>
+<title>ViewPortletDMDShowSearchTest</title>
 </head>
 <body>
 <table cellpadding="1" cellspacing="1" border="1">
 <thead>
-<tr><td rowspan="1" colspan="3">ViewPortletShowSearchTest</td></tr>
+<tr><td rowspan="1" colspan="3">ViewPortletDMDShowSearchTest</td></tr>
 </thead><tbody>
 <tr>
 	<td>open</td>
